Scripts run_mcbeds and update_server have been combined and completely rewritten into one monolithic script, mcbeds_mgr, which also includes additional features.
Options previously defined in the head of the run_mcbeds script have been
moved into .mcbedsrc, and include additional options. This file can also be
stored for system-wide use in /etc/mcbeds.rc or /usr/local/etc/mcbeds.rc

mcbeds_mgr can be called as mcbeds_start to start the server, mcbeds_stop to
stop the server, and mcbeds_update to update the server. The script,
regardless of how it is called, can also be controlled with flags to override
default or define new behaviors.

Updated README.md to reflect these changes